app/requirements.txt
aiohttp==3.9.5
aioice==0.9.0
aiortc==1.8.0
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
av==11.0.0
bidict==0.23.1
cffi==1.16.0
click==8.1.7
cryptography==42.0.7
dnspython==2.6.1
frozenlist==1.4.1
google-crc32c==1.5.0
h11==0.14.0
idna==3.7
ifaddr==0.2.0
linkify-it-py==2.0.3
markdown-it-py==3.0.0
mdit-py-plugins==0.4.1
mdurl==0.1.2
msgpack==1.0.8
multidict==6.0.5
pycparser==2.22
pyee==11.1.0
Pygments==2.18.0
pylibsrtp==0.10.0
pyOpenSSL==24.1.0
python-engineio==4.3.1
python-socketio==5.5.1
rich==13.7.1
simple-websocket==1.0.0
textual==0.63.6
textual-dev==1.5.1
typing_extensions==4.12.0
uc-micro-py==1.0.3
wsproto==1.2.0
yarl==1.9.4
